Add latin_error.py
This is a snippet of Python that fixed a common error in our internal software. If a data analyst did not use the correct Ruby function to create a .csv file of their work (that function being write_csv, which automatically encodes everything in UTF-8) then sometimes certain files would use latin-1 instead. This would cause a UnicodeError in Python when our software tried to read the file.

This snippet fixed that by trying to read it using latin-1, if/when it through a UnicodeError.

(Print statements still included, they were used for testing this snippet.)